backend/requirements.txt
fastapi
uvicorn
pandas
scikit-learn
python-dotenv
sqlalchemy
psycopg2-binary